[BREAKING] Use u32 for bit positions
usize is larger than we need, and things like `BITS` constants
associated with primitive integral types are already u32.

We update the version to 0.4.

+++ src/lib.rs
@@ -454,9 +454,9 @@ pub struct FieldMetadata<Base: Unsigned> {
     /// The name of the bitfield.
     pub name: &'static str,
     /// The high bit of the bitfield.
-    pub high_bit: usize,
+    pub high_bit: u32,
     /// The low bit of the bitfield.
-    pub low_bit: usize,
+    pub low_bit: u32,
     /// The default value of the bitfield.
     pub default: Base,
 }
